refactor all classes into diffrent dirs
