+++ models/losses/__init__.py
@@ -1,0 +1,8 @@
+from .base_class import BasePINNLoss
+from .pinn import PINN_Loss
+from .sapinn import SA_PINN_Loss
+from .dbpinn import DB_PINN_Loss
+
+__all__ = [
+    "BasePINNLoss", "PINN_Loss", "SA_PINN_Loss", "DB_PINN_Loss"
+]
